S08P31A506-230 [feat] 광고 이벤트 처리 Repository 구현

+++ backend/statistics/src/main/java/com/avocado/statistics/db/redis/repository/MerchandiseIdSetRepository.java
@@ -20,7 +20,7 @@ public class MerchandiseIdSetRepository {
         this.bitSet = redisson.getBitSet(key);
     }
 
-    public void setUse(int merchandiseId) {
+    public void setUse(Long merchandiseId) {
         bitSet.set(merchandiseId);
     }
